ajoute functionaliter delete client

+++ src/ui/ClientView.java
@@ -1,20 +1,26 @@
 package ui;
 
 import dao.ClientDAO;
+import entity.Carte;
 import entity.Client;
+import service.CarteService;
 import service.ClientService;
 
+import java.util.List;
 import java.util.Scanner;
 
 public class ClientView {
     private final ClientService clientService;
+    private CarteService carteService;
+    private final MenuUI menu;
 
-
-    public ClientView(ClientService clientService) {
+    public ClientView(ClientService clientService, CarteService carteService, MenuUI menu) {
+        this.carteService = carteService;
         this.clientService = clientService;
+        this.menu = menu;
     }
 
-    public void ajouteClient(){
+    public void ajouteClient() {
 
         Scanner scanner = new Scanner(System.in);
         try {
@@ -26,21 +32,31 @@ public class ClientView {
             String email = scanner.nextLine();
             System.out.println("Entrer le téléphone du client:");
             String telephone = scanner.nextLine();
-            Client client = new Client(id,nom,email,telephone);
+            Client client = new Client(id, nom, email, telephone);
             clientService.createClient(client);
             System.out.println("Client ajouté avec succès !");
         } catch (IllegalArgumentException e) {
             System.out.println(e.getMessage());
         }
 
-
     }
 
-    public void deleteClient(){
-        Scanner scanner = new Scanner(System.in);
+    public void deleteClient() {
         try {
-            System.out.println("Entrer l'ID du client à supprimer:");
-            int id = Integer.parseInt(scanner.nextLine());
+            int id = menu.saisirEntierSecurise("Entrer l'ID du client à supprimer: ");
+
+            if (id == -1) {
+                System.out.println("Opération annulée - ID invalide");
+                return;
+            }
+
+            List<Carte> cartes = carteService.listerCartesParClient(id);
+
+            if (!cartes.isEmpty()) {
+                System.out.println("Impossible de supprimer : ce client possède " + cartes.size() + " carte(s)");
+                return;
+            }
+
             clientService.deleteClient(id);
             System.out.println("Client supprimé avec succès !");
         } catch (IllegalArgumentException e) {
